rgeis26-assignment-10/requirements.txt
flask
torch
open-clip-torch
pandas
Pillow
matplotlib
scikit-learn
